+++ portage-ucp/lib/portage/ucp/capabilities/order.rb
@@ -4,7 +4,12 @@ module Portage
       ORDER = Portage::Ucp::Capability.new(
         name: "dev.ucp.shopping.order",
         version: "1",
-        actions: { "get_order" => :get_order }
+        actions: {
+          "get_order" => :get_order,
+          "cancel_order" => :cancel_order,
+          "request_return" => :request_return,
+          "refund_order" => :refund_order
+        }
       )
     end
   end